Improvements to RTDS provider
- Remove PMU sequence numbers.
They ended up being wildly inaccurate thanks to threading.
- Apply a field type mapping when creating Elasticsearch indices
- Handle Elastic bulk ingest error messages
- Add 'elastic-index-basename' option to configure
the Elasticsearch index name.
- Fix PMU connection rebuilding.
If RSCAD case is stopped, the provider no longer needs to be restarted and will properly restart the PMU connections.
This fixes a long-standing issue that's been around since HARMONIE.
- Improve PMU handling.
RTDS provider is now able to be started AFTER the RSCAD case is started.
- Change gtnet-skt to continually send data over UDP
- Add 'gtnet-skt-tcp-retry-delay' config option
- Add 'gtnet-skt-udp-write-rate' config option
- Move Elasticsearch processing and pushing into a thread.
This prevents it from blocking the core functionality of
sending data to SCEPTRE.
- Improve how boolean config options are handled
- Move RTDS data processing into a thread